Add recommended EEROS version and exact check
Set v1.0.0.0 as recommended EEROS Framework version.
Add logic to check if version information is set already
in CMake Cache. If it is set already and does not match
the recommended version, a message will be printed.
If it is not set, the recommended version will be used instead.

The EEROS version must match exactly.